Complete updates for Elastic's API v5.6.10
This is based on the previous work on this branch and adds the following
changes:

- Integration tests now inherit from ESIntegTestCase in order to be
compatible with the randomized context that it's running in. Avoids
errors like `failed to invoke RandomizedContext#getRandom`
- If the balancer doesn't assign all replicas, it now explicitly marks
them as ignored for this run instead of leaving them in the list of
unassigned shards in the `RoutingNodes`. If it didn't do that, Elastic
would complain about duplicated shards in the routing table and merrily
throw exceptions.
- When upgrading the elasticsearch-cluster-runner, we had to exclude the
`org.aw2.ssm:asm-debug-all` dependency because of a version conflict.
This only affects the test scope anyway.
- All plugins must now use at least Java 8, so that's now reflected in
the POM and the plugin description.
- When serialising JSON responses in `ActionResponse`s, we now need to
explicitly open the root object.
- We can't rely on `createComponents()` being called for the plugin all
the time, so I added a lazy initialiser. This might only be needed for
the test setup, though.
- The elasticsearch-cluster-runner plugin has decided that it wants to
be told about which plugins to install in a different way now. That's
also addressed.
- `ClusterInfo` now takes immutable maps for things like the shard
sizes, so it's a bit more tricky to mock. That's done now, too.